requirements.txt
#Run
requests
PyYAML
progressbar
